Easily work with multiple datasets
(In theory, not tested.) Have a config file with all the paths that both
the Python scripts and the Bash scripts load. Added in the Sloth
labeling scripts and a script to convert from the JSON to YOLO's format.
Added script to compress only images we have labels for and will use in
training, which is all we'll upload to Kamiak.