Limit positions for TensorFlow to be in [0,1]
Otherwise it errors if it's just a little bit out of the image. Also,
was syncing the wrong directory for log files. Now it'll sync the TF log
files and allow you to see the results in TensorBoard. Nice.
Unfortunately, tf_eval.srun doesn't work because it needs Matplotlib
which requires _tkinter. So, evaluation doesn't work.