Initial attempt at getting 3D point for bounding boxes
Using YOLO at the moment. Gives us an (x,y,z) point for the center
of each bounding box. However, not sure what this is referenced from.
Also, should not just use center since if you look at the 3D point cloud
in rviz, the center is likely not actually at the object, e.g. with the
pillbottle the center is on the table behind the pillbottle.