Include COCO human images in training data
This in turns mean we have to deal with the large class imbalance. I did
this through duplicating training examples and throwing in more data
augmentation randomness. I looked at SMOTE but it looked too hard to set
up quickly and it might not work well for images -- not sure.